+++ rpc/src/v1/tests/helpers/sync_provider.rs
@@ -59,11 +59,5 @@ impl SyncProvider for TestSyncProvider {
 	fn status(&self) -> SyncStatus {
 		self.status.read().unwrap().clone()
 	}
-
-	fn start_network(&self) {
-	}
-
-	fn stop_network(&self) {
-	}
 }
