+++ rpc/src/v1/traits/net.rs
@@ -30,12 +30,6 @@ pub trait Net: Sized + Send + Sync + 'static {
 	/// Otherwise false.
 	fn is_listening(&self, _: Params) -> Result<Value, Error>;
 
-	/// Start the network.
-	fn start_network(&self, _: Params) -> Result<Value, Error>;
-
-	/// Stop the network.
-	fn stop_network(&self, _: Params) -> Result<Value, Error>;
-
 	/// Should be used to convert object to io delegate.
 	fn to_delegate(self) -> IoDelegate<Self> {
 		let mut delegate = IoDelegate::new(Arc::new(self));
